requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
PyPDF2==3.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
python-docx==0.8.11
pillow==10.0.1
matplotlib==3.7.2
plotly==5.16.1
pandas==2.0.3
numpy==1.24.3